server/requirements.txt
Django==2.2
djangorestframework==3.9.2
pytz==2018.9
sqlparse==0.3.0
django-enumfields==1.0.0
djangorestframework-jwt==1.11.0
PyJWT==1.6.1